feat: show loader in embed mode
