+++ src/components/Loader.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import frog from 'url:~/public/128-production.png';
+
+function Loader({ loading }) {
+  return (
+    <div
+      className={[
+        'w-full h-full absolute top-0 left-0 flex flex-col justify-center items-center w-full h-full space-y-4 fade',
+        loading ? 'opacity-100' : 'opacity-0',
+      ].join(' ')}
+    >
+      <img className="opacity-50" src={frog} />
+    </div>
+  );
+}
+
+export default Loader;
