+++ src/components/Playground.js
@@ -6,20 +6,7 @@ import Result from './Result';
 import Query from './Query';
 import usePlayground from '../hooks/usePlayground';
 import Layout from './Layout';
-import frog from 'url:~/public/128-production.png';
-
-function Loader({ loading }) {
-  return (
-    <div
-      className={[
-        'w-full h-full absolute top-0 left-0 flex flex-col justify-center items-center w-full h-full space-y-4 fade',
-        loading ? 'opacity-100' : 'opacity-0',
-      ].join(' ')}
-    >
-      <img className="opacity-50" src={frog} />
-    </div>
-  );
-}
+import Loader from './Loader';
 
 function Paper({ children }) {
   return (
